refactor into a few more modules
